+++ wcfsetup/install/files/lib/data/comment/ViewableCommentList.class.php
@@ -1,5 +1,6 @@
 <?php
 namespace wcf\data\comment;
+use wcf\data\user\UserProfileCache;
 
 /**
  * Represents a list of decorated comment objects.
@@ -17,16 +18,20 @@ class ViewableCommentList extends CommentList {
 	 */
 	public $decoratorClassName = 'wcf\data\comment\ViewableComment';
 	
-	/**
-	 * Creates a new ViewableCommentList object.
-	 */
-	public function __construct() {
-		parent::__construct();
+	public function readObjects() {
+		parent::readObjects();
 		
-		// get avatars
-		if (!empty($this->sqlSelects)) $this->sqlSelects .= ',';
-		$this->sqlSelects .= "user_avatar.*, user_table.*";
-		$this->sqlJoins .= " LEFT JOIN wcf".WCF_N."_user user_table ON (user_table.userID = comment.userID)";
-		$this->sqlJoins .= " LEFT JOIN wcf".WCF_N."_user_avatar user_avatar ON (user_avatar.avatarID = user_table.avatarID)";
+		if (!empty($this->objects)) {
+			$userIDs = array();
+			foreach ($this->objects as $comment) {
+				if ($comment->userID) {
+					$userIDs[] = $comment->userID;
+				}
+			}
+			
+			if (!empty($userIDs)) {
+				UserProfileCache::getInstance()->cacheUserIDs($userIDs);
+			}
+		}
 	}
 }