+++ wcfsetup/install/files/lib/system/user/activity/event/ProfileCommentResponseUserActivityEvent.class.php
@@ -2,7 +2,7 @@
 namespace wcf\system\user\activity\event;
 use wcf\data\comment\response\CommentResponseList;
 use wcf\data\comment\CommentList;
-use wcf\data\user\UserProfileList;
+use wcf\data\user\UserProfileCache;
 use wcf\system\user\activity\event\IUserActivityEvent;
 use wcf\system\SingletonFactory;
 use wcf\system\WCF;
@@ -56,10 +56,7 @@ class ProfileCommentResponseUserActivityEvent extends SingletonFactory implement
 			$userIDs[] = $comment->userID;
 		}
 		if (!empty($userIDs)) {
-			$userList = new UserProfileList();
-			$userList->setObjectIDs($userIDs);
-			$userList->readObjects();
-			$users = $userList->getObjects();
+			$users = UserProfileCache::getInstance()->getUserProfiles($userIDs);
 		}
 		
 		// set message